Unify Spotlight default + draggable experiments
