+++ samples/sandbox2/src/main/java/com/bumble/appyx/sandbox2/MainActivity.kt
@@ -5,7 +5,6 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
@@ -30,7 +29,6 @@ import com.bumble.appyx.sandbox2.navmodel2.BackStackExperimentDebug
 import com.bumble.appyx.sandbox2.navmodel2.PromoterExperiment
 import com.bumble.appyx.sandbox2.navmodel2.SpotlightExperiment
 import com.bumble.appyx.sandbox2.navmodel2.SpotlightExperimentDebug
-import com.bumble.appyx.sandbox2.navmodel2.SpotlightExperimentDrag
 import com.bumble.appyx.sandbox2.ui.theme.AppyxTheme
 
 @ExperimentalMaterialApi
@@ -58,14 +56,12 @@ class MainActivity : ComponentActivity() {
                             Button({ content = 1 }) { Text("2") }
                             Button({ content = 2 }) { Text("3") }
                             Button({ content = 3 }) { Text("4") }
-                            Button({ content = 4 }) { Text("5") }
                         }
                         when (content) {
                             0 -> PromoterExperiment()
                             1 -> SpotlightExperiment()
                             2 -> SpotlightExperimentDebug()
-                            3 -> SpotlightExperimentDrag()
-                            4 -> BackStackExperimentDebug()
+                            3 -> BackStackExperimentDebug()
                             else -> SpotlightExperiment()
                         }
                     }
