+++ samples/sandbox2/src/main/java/com/bumble/appyx/sandbox2/navmodel2/SpotlightExperiment.kt
@@ -1,8 +1,11 @@
 package com.bumble.appyx.sandbox2.navmodel2
 
+import android.util.Log
 import androidx.compose.animation.core.Spring
 import androidx.compose.animation.core.spring
+import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.Orientation
+import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -20,9 +23,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
 import com.bumble.appyx.core.navigation2.inputsource.AnimatedInputSource
+import com.bumble.appyx.core.navigation2.inputsource.DragProgressInputSource
 import com.bumble.appyx.navmodel2.spotlight.Spotlight
 import com.bumble.appyx.navmodel2.spotlight.operation.first
 import com.bumble.appyx.navmodel2.spotlight.operation.last
@@ -36,43 +42,73 @@ import com.bumble.appyx.sandbox2.navmodel2.NavTarget.Child4
 import com.bumble.appyx.sandbox2.navmodel2.NavTarget.Child5
 import com.bumble.appyx.sandbox2.navmodel2.NavTarget.Child6
 import com.bumble.appyx.sandbox2.navmodel2.NavTarget.Child7
+import com.bumble.appyx.sandbox2.ui.theme.appyx_dark
 import kotlinx.coroutines.flow.map
-
-
-private val spotlight = Spotlight(
-    items = listOf(Child1, Child2, Child3, Child4, Child5, Child6, Child7),
-)
 
 
 @ExperimentalMaterialApi
 @Composable
 fun SpotlightExperiment() {
+    val spotlight = remember {
+        Spotlight(
+            items = listOf(Child1, Child2, Child3, Child4, Child5, Child6, Child7),
+        )
+    }
     val coroutineScope = rememberCoroutineScope()
-    val inputSource = remember {
+    val drag = remember { DragProgressInputSource(spotlight, coroutineScope) }
+    val animated = remember {
         AnimatedInputSource(
-            navModel = spotlight,
-            coroutineScope = coroutineScope,
-            defaultAnimationSpec = spring(
+            spotlight, coroutineScope, spring(
                 stiffness = Spring.StiffnessMediumLow
             )
         )
     }
 
+    val density = LocalDensity.current
     var elementSize by remember { mutableStateOf(IntSize(0, 0)) }
     val transitionParams by createTransitionParams(elementSize)
-    val uiProps = remember(transitionParams) { SpotlightSlider<NavTarget>(
-        transitionParams = transitionParams,
-        orientation = Orientation.Horizontal
-    ) }
+    val uiProps = remember(transitionParams) {
+        SpotlightSlider<NavTarget>(
+            transitionParams = transitionParams,
+            orientation = Orientation.Horizontal
+        )
+    }
     val render = remember(uiProps) { spotlight.segments.map { uiProps.map(it) } }
 
     Column(
-        modifier = Modifier.fillMaxSize()
+        Modifier
+            .fillMaxWidth()
+            .background(appyx_dark)
     ) {
         Children(
             render = render.collectAsState(listOf()),
             modifier = Modifier.weight(0.9f),
             onElementSizeChanged = { elementSize = it },
+            element = {
+                Element(
+                    render = it,
+                    modifier = Modifier
+                        .fillMaxSize()
+                        .pointerInput(it.navElement.key) {
+                            detectDragGestures(
+                                onDrag = { change, dragAmount ->
+                                    change.consume()
+                                    if (drag.gestureFactory == null) {
+                                        drag.gestureFactory = { dragAmount ->
+                                            uiProps.createGesture(dragAmount, density)
+                                        }
+                                    }
+                                    drag.addDeltaProgress(dragAmount)
+                                },
+                                onDragEnd = {
+                                    Log.d("drag", "end")
+                                    drag.gestureFactory = null
+                                    drag.settle()
+                                }
+                            )
+                        }
+                )
+            }
         )
 
         Row(
@@ -82,18 +118,19 @@ fun SpotlightExperiment() {
                 .padding(4.dp),
             horizontalArrangement = Arrangement.SpaceEvenly
         ) {
-            Button(onClick = { inputSource.first() }) {
+            Button(onClick = { animated.first() }) {
                 Text("First")
             }
-            Button(onClick = { inputSource.previous(spring(stiffness = Spring.StiffnessLow)) }) {
+            Button(onClick = { animated.previous(spring(stiffness = Spring.StiffnessLow)) }) {
                 Text("Prev")
             }
-            Button(onClick = { inputSource.next(spring(stiffness = Spring.StiffnessMedium)) }) {
+            Button(onClick = { animated.next(spring(stiffness = Spring.StiffnessMedium)) }) {
                 Text("Next")
             }
-            Button(onClick = { inputSource.last() }) {
+            Button(onClick = { animated.last() }) {
                 Text("Last")
             }
         }
     }
 }
+
